# artifacts/email-assistant/pyapi-openapi.json
{"openapi":"3.1.0","info":{"title":"Email Assistant API","description":"Personalized email assistant with classification, draft generation, and RL feedback loop","version":"1.0.0"},"paths":{"/emails":{"post":{"tags":["emails"],"summary":"Create Email","operationId":"create_email_emails_post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmailCreateIn"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmailOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["emails"],"summary":"List Emails","operationId":"list_emails_emails_get","parameters":[{"name":"priority","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Priority"}},{"name":"intent","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Intent"}},{"name":"action","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Action"}},{"name":"is_read","in":"query","required":false,"schema":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Is Read"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":100,"default":50,"title":"Limit"}},{"name":"offset","in":"query","required":false,"schema":{"type":"integer","minimum":0,"default":0,"title":"Offset"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmailList"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/emails/{email_id}/classify":{"post":{"tags":["emails"],"summary":"Reclassify Email","operationId":"reclassify_email_emails__email_id__classify_post","parameters":[{"name":"email_id","in":"path","required":true,"schema":{"type":"integer","title":"Email Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmailOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/emails/{email_id}":{"get":{"tags":["emails"],"summary":"Get Email","operationId":"get_email_emails__email_id__get","parameters":[{"name":"email_id","in":"path","required":true,"schema":{"type":"integer","title":"Email Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmailOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/emails/{email_id}/draft":{"post":{"tags":["drafts"],"summary":"Generate Email Draft","operationId":"generate_email_draft_emails__email_id__draft_post","parameters":[{"name":"email_id","in":"path","required":true,"schema":{"type":"integer","title":"Email Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DraftRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DraftOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/emails/{email_id}/feedback":{"post":{"tags":["feedback"],"summary":"Submit Feedback","operationId":"submit_feedback_emails__email_id__feedback_post","parameters":[{"name":"email_id","in":"path","required":true,"schema":{"type":"integer","title":"Email Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FeedbackIn"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FeedbackOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/feedback":{"get":{"tags":["feedback"],"summary":"List Feedback","operationId":"list_feedback_feedback_get","parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":50,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/FeedbackOut"},"title":"Response List Feedback Feedback Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/preferences":{"get":{"tags":["preferences"],"summary":"List Preferences","operationId":"list_preferences_preferences_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/SenderPreferenceOut"},"type":"array","title":"Response List Preferences Preferences Get"}}}}}},"post":{"tags":["preferences"],"summary":"Create Preference","operationId":"create_preference_preferences_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SenderPreferenceIn"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SenderPreferenceOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/preferences/{pref_id}":{"put":{"tags":["preferences"],"summary":"Update Preference","operationId":"update_preference_preferences__pref_id__put","parameters":[{"name":"pref_id","in":"path","required":true,"schema":{"type":"integer","title":"Pref Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SenderPreferenceIn"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SenderPreferenceOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["preferences"],"summary":"Delete Preference","operationId":"delete_preference_preferences__pref_id__delete","parameters":[{"name":"pref_id","in":"path","required":true,"schema":{"type":"integer","title":"Pref Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/tone-rules":{"get":{"tags":["tone-rules"],"summary":"List Tone Rules","operationId":"list_tone_rules_tone_rules_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/ToneRuleOut"},"type":"array","title":"Response List Tone Rules Tone Rules Get"}}}}}},"post":{"tags":["tone-rules"],"summary":"Create Tone Rule","operationId":"create_tone_rule_tone_rules_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ToneRuleIn"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ToneRuleOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/tone-rules/{rule_id}":{"put":{"tags":["tone-rules"],"summary":"Update Tone Rule","operationId":"update_tone_rule_tone_rules__rule_id__put","parameters":[{"name":"rule_id","in":"path","required":true,"schema":{"type":"integer","title":"Rule Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ToneRuleIn"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ToneRuleOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["tone-rules"],"summary":"Delete Tone Rule","operationId":"delete_tone_rule_tone_rules__rule_id__delete","parameters":[{"name":"rule_id","in":"path","required":true,"schema":{"type":"integer","title":"Rule Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/analytics":{"get":{"tags":["analytics"],"summary":"Get Analytics","operationId":"get_analytics_analytics_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AnalyticsOut"}}}}}}},"/history":{"get":{"tags":["analytics"],"summary":"Get History","operationId":"get_history_history_get","parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":50,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/HistoryItem"},"title":"Response Get History History Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/seed":{"post":{"tags":["seed"],"summary":"Seed Database","operationId":"seed_database_seed_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/rl/settings":{"get":{"tags":["rl"],"summary":"Get Rl Settings","operationId":"get_rl_settings_rl_settings_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RLSettings"}}}}}},"put":{"tags":["rl"],"summary":"Set Rl Model","operationId":"set_rl_model_rl_settings_put","parameters":[{"name":"model_key","in":"query","required":true,"schema":{"type":"string","description":"Model key: 'qwen' or 'gemma'","title":"Model Key"},"description":"Model key: 'qwen' or 'gemma'"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/healthz":{"get":{"summary":"Health","operationId":"health_healthz_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"ActionBreakdownItem":{"properties":{"action":{"type":"string","title":"Action"},"count":{"type":"integer","title":"Count"},"percentage":{"type":"number","title":"Percentage"}},"type":"object","required":["action","count","percentage"],"title":"ActionBreakdownItem"},"AnalyticsOut":{"properties":{"reward_trend":{"items":{"$ref":"#/components/schemas/RewardTrendPoint"},"type":"array","title":"Reward Trend"},"tone_distribution":{"items":{"$ref":"#/components/schemas/ToneDistributionItem"},"type":"array","title":"Tone Distribution"},"action_breakdown":{"items":{"$ref":"#/components/schemas/ActionBreakdownItem"},"type":"array","title":"Action Breakdown"},"bandit_state":{"items":{"$ref":"#/components/schemas/BanditState"},"type":"array","title":"Bandit State"},"priority_breakdown":{"items":{"$ref":"#/components/schemas/ClassificationBreakdownItem"},"type":"array","title":"Priority Breakdown"},"intent_breakdown":{"items":{"$ref":"#/components/schemas/ClassificationBreakdownItem"},"type":"array","title":"Intent Breakdown"},"rl_buffer":{"$ref":"#/components/schemas/RLBufferStats"},"rl_model_breakdown":{"items":{"$ref":"#/components/schemas/RLModelBreakdownItem"},"type":"array","title":"Rl Model Breakdown"},"total_emails_processed":{"type":"integer","title":"Total Emails Processed"},"avg_overall_reward":{"type":"number","title":"Avg Overall Reward"}},"type":"object","required":["reward_trend","tone_distribution","action_breakdown","bandit_state","priority_breakdown","intent_breakdown","rl_buffer","rl_model_breakdown","total_emails_processed","avg_overall_reward"],"title":"AnalyticsOut"},"BanditState":{"properties":{"tone":{"type":"string","title":"Tone"},"avg_reward":{"type":"number","title":"Avg Reward"},"count":{"type":"integer","title":"Count"},"expected_reward":{"type":"number","title":"Expected Reward"}},"type":"object","required":["tone","avg_reward","count","expected_reward"],"title":"BanditState"},"ClassificationBreakdownItem":{"properties":{"label":{"type":"string","title":"Label"},"count":{"type":"integer","title":"Count"},"percentage":{"type":"number","title":"Percentage"}},"type":"object","required":["label","count","percentage"],"title":"ClassificationBreakdownItem"},"DraftOut":{"properties":{"email_id":{"type":"integer","title":"Email Id"},"draft":{"type":"string","title":"Draft"},"tone_used":{"type":"string","title":"Tone Used"},"style_notes":{"type":"string","title":"Style Notes"},"similar_past_emails":{"items":{"$ref":"#/components/schemas/SimilarEmail"},"type":"array","title":"Similar Past Emails"}},"type":"object","required":["email_id","draft","tone_used","style_notes","similar_past_emails"],"title":"DraftOut"},"DraftRequest":{"properties":{"tone_override":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Tone Override"}},"type":"object","title":"DraftRequest"},"EmailCreateIn":{"properties":{"subject":{"type":"string","title":"Subject"},"sender_name":{"type":"string","title":"Sender Name"},"sender_email":{"type":"string","title":"Sender Email"},"body":{"type":"string","title":"Body"},"received_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Received At"}},"type":"object","required":["subject","sender_name","sender_email","body"],"title":"EmailCreateIn"},"EmailList":{"properties":{"emails":{"items":{"$ref":"#/components/schemas/EmailOut"},"type":"array","title":"Emails"},"total":{"type":"integer","title":"Total"}},"type":"object","required":["emails","total"],"title":"EmailList"},"EmailOut":{"properties":{"id":{"type":"integer","title":"Id"},"subject":{"type":"string","title":"Subject"},"sender_name":{"type":"string","title":"Sender Name"},"sender_email":{"type":"string","title":"Sender Email"},"body":{"type":"string","title":"Body"},"received_at":{"type":"string","format":"date-time","title":"Received At"},"is_read":{"type":"boolean","title":"Is Read"},"priority":{"type":"string","title":"Priority"},"intent":{"type":"string","title":"Intent"},"recommended_action":{"type":"string","title":"Recommended Action"},"classification_confidence":{"type":"number","title":"Classification Confidence"},"classification_notes":{"type":"string","title":"Classification Notes"},"rl_model":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Rl Model"},"rl_model_key":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Rl Model Key"},"rl_positive_examples":{"type":"integer","title":"Rl Positive Examples","default":0},"rl_negative_examples":{"type":"integer","title":"Rl Negative Examples","default":0},"rl_latency_ms":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Rl Latency Ms"},"rl_active":{"type":"boolean","title":"Rl Active","default":false},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","subject","sender_name","sender_email","body","received_at","is_read","priority","intent","recommended_action","classification_confidence","classification_notes","created_at"],"title":"EmailOut"},"FeedbackIn":{"properties":{"user_action":{"type":"string","title":"User Action"},"final_output":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Final Output"},"tone_used":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Tone Used"},"edit_distance":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Edit Distance","default":0}},"type":"object","required":["user_action"],"title":"FeedbackIn"},"FeedbackOut":{"properties":{"id":{"type":"integer","title":"Id"},"email_id":{"type":"integer","title":"Email Id"},"user_action":{"type":"string","title":"User Action"},"reward":{"type":"number","title":"Reward"},"tone_used":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Tone Used"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","email_id","user_action","reward","tone_used","created_at"],"title":"FeedbackOut"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"HistoryItem":{"properties":{"id":{"type":"integer","title":"Id"},"email_id":{"type":"integer","title":"Email Id"},"user_action":{"type":"string","title":"User Action"},"reward":{"type":"number","title":"Reward"},"tone_used":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Tone Used"},"edit_distance":{"type":"integer","title":"Edit Distance"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"email_subject":{"type":"string","title":"Email Subject"},"email_sender":{"type":"string","title":"Email Sender"}},"type":"object","required":["id","email_id","user_action","reward","tone_used","edit_distance","created_at","email_subject","email_sender"],"title":"HistoryItem"},"RLBufferStats":{"properties":{"total_experiences":{"type":"integer","title":"Total Experiences"},"positive_experiences":{"type":"integer","title":"Positive Experiences"},"negative_experiences":{"type":"integer","title":"Negative Experiences"},"neutral_experiences":{"type":"integer","title":"Neutral Experiences"},"positive_ratio":{"type":"number","title":"Positive Ratio"},"avg_reward":{"type":"number","title":"Avg Reward"},"recent_avg_reward":{"type":"number","title":"Recent Avg Reward"},"learning_active":{"type":"boolean","title":"Learning Active"}},"type":"object","required":["total_experiences","positive_experiences","negative_experiences","neutral_experiences","positive_ratio","avg_reward","recent_avg_reward","learning_active"],"title":"RLBufferStats"},"RLModelBreakdownItem":{"properties":{"model_key":{"type":"string","title":"Model Key"},"model_id":{"type":"string","title":"Model Id"},"count":{"type":"integer","title":"Count"},"avg_confidence":{"type":"number","title":"Avg Confidence"}},"type":"object","required":["model_key","model_id","count","avg_confidence"],"title":"RLModelBreakdownItem"},"RLSettings":{"properties":{"model_key":{"type":"string","title":"Model Key"},"model_id":{"type":"string","title":"Model Id"},"available_models":{"additionalProperties":true,"type":"object","title":"Available Models"}},"type":"object","required":["model_key","model_id","available_models"],"title":"RLSettings"},"RewardTrendPoint":{"properties":{"date":{"type":"string","title":"Date"},"avg_reward":{"type":"number","title":"Avg Reward"},"count":{"type":"integer","title":"Count"}},"type":"object","required":["date","avg_reward","count"],"title":"RewardTrendPoint"},"SenderPreferenceIn":{"properties":{"sender_email":{"type":"string","title":"Sender Email"},"sender_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Sender Name"},"importance":{"type":"string","title":"Importance","default":"medium"},"default_tone":{"type":"string","title":"Default Tone","default":"professional"},"default_action":{"type":"string","title":"Default Action","default":"reply"},"reply_priority":{"type":"string","title":"Reply Priority","default":"medium"},"notes":{"type":"string","title":"Notes","default":""}},"type":"object","required":["sender_email"],"title":"SenderPreferenceIn"},"SenderPreferenceOut":{"properties":{"id":{"type":"integer","title":"Id"},"sender_email":{"type":"string","title":"Sender Email"},"sender_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Sender Name"},"importance":{"type":"string","title":"Importance"},"default_tone":{"type":"string","title":"Default Tone"},"default_action":{"type":"string","title":"Default Action"},"reply_priority":{"type":"string","title":"Reply Priority"},"notes":{"type":"string","title":"Notes"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","required":["id","sender_email","sender_name","importance","default_tone","default_action","reply_priority","notes","created_at","updated_at"],"title":"SenderPreferenceOut"},"SimilarEmail":{"properties":{"subject":{"type":"string","title":"Subject"},"reply":{"type":"string","title":"Reply"}},"type":"object","required":["subject","reply"],"title":"SimilarEmail"},"ToneDistributionItem":{"properties":{"tone":{"type":"string","title":"Tone"},"count":{"type":"integer","title":"Count"},"avg_reward":{"type":"number","title":"Avg Reward"}},"type":"object","required":["tone","count","avg_reward"],"title":"ToneDistributionItem"},"ToneRuleIn":{"properties":{"name":{"type":"string","title":"Name"},"rule_type":{"type":"string","title":"Rule Type"},"value":{"type":"string","title":"Value"},"is_active":{"type":"boolean","title":"Is Active","default":true}},"type":"object","required":["name","rule_type","value"],"title":"ToneRuleIn"},"ToneRuleOut":{"properties":{"id":{"type":"integer","title":"Id"},"name":{"type":"string","title":"Name"},"rule_type":{"type":"string","title":"Rule Type"},"value":{"type":"string","title":"Value"},"is_active":{"type":"boolean","title":"Is Active"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","name","rule_type","value","is_active","created_at"],"title":"ToneRuleOut"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}