# artifacts/email-assistant/src/generated/pyapi/model/analyticsOut.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Email Assistant API
 * Personalized email assistant with classification, draft generation, and RL feedback loop
 * OpenAPI spec version: 1.0.0
 */
import type { ActionBreakdownItem } from './actionBreakdownItem';
import type { BanditState } from './banditState';
import type { ClassificationBreakdownItem } from './classificationBreakdownItem';
import type { RewardTrendPoint } from './rewardTrendPoint';
import type { RLBufferStats } from './rLBufferStats';
import type { RLModelBreakdownItem } from './rLModelBreakdownItem';
import type { ToneDistributionItem } from './toneDistributionItem';

export type AnalyticsOut = {
  reward_trend: RewardTrendPoint[];
  tone_distribution: ToneDistributionItem[];
  action_breakdown: ActionBreakdownItem[];
  bandit_state: BanditState[];
  priority_breakdown: ClassificationBreakdownItem[];
  intent_breakdown: ClassificationBreakdownItem[];
  rl_buffer: RLBufferStats;
  rl_model_breakdown: RLModelBreakdownItem[];
  total_emails_processed: number;
  avg_overall_reward: number;
};