# artifacts/email-assistant/src/generated/pyapi/model/emailOut.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Email Assistant API
 * Personalized email assistant with classification, draft generation, and RL feedback loop
 * OpenAPI spec version: 1.0.0
 */

export type EmailOut = {
  id: number;
  subject: string;
  sender_name: string;
  sender_email: string;
  body: string;
  received_at: string;
  is_read: boolean;
  priority: string;
  intent: string;
  recommended_action: string;
  classification_confidence: number;
  classification_notes: string;
  rl_model?: string | null;
  rl_model_key?: string | null;
  rl_positive_examples?: number;
  rl_negative_examples?: number;
  rl_latency_ms?: number | null;
  rl_active?: boolean;
  created_at: string;
};