# artifacts/email-assistant/src/generated/pyapi/model/rLBufferStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Email Assistant API
 * Personalized email assistant with classification, draft generation, and RL feedback loop
 * OpenAPI spec version: 1.0.0
 */

export type RLBufferStats = {
  total_experiences: number;
  positive_experiences: number;
  negative_experiences: number;
  neutral_experiences: number;
  positive_ratio: number;
  avg_reward: number;
  recent_avg_reward: number;
  learning_active: boolean;
};