# artifacts/email-assistant/src/generated/pyapi/model/rLModelBreakdownItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Email Assistant API
 * Personalized email assistant with classification, draft generation, and RL feedback loop
 * OpenAPI spec version: 1.0.0
 */

export type RLModelBreakdownItem = {
  model_key: string;
  model_id: string;
  count: number;
  avg_confidence: number;
};