# artifacts/email-assistant/src/generated/pyapi/model/rLSettings.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Email Assistant API
 * Personalized email assistant with classification, draft generation, and RL feedback loop
 * OpenAPI spec version: 1.0.0
 */
import type { RLSettingsAvailableModels } from './rLSettingsAvailableModels';

export type RLSettings = {
  model_key: string;
  model_id: string;
  available_models: RLSettingsAvailableModels;
};